refactor: Address review comments on testing infrastructure
- Rename 'cucumber' profile to 'bdd' (behavior-driven development)
- Fix manual-entry.feature: change role from "administrator" to "investor"
- Add documentation explaining why Background step definitions are empty
- Rename 'fixtures' package to 'testutils' for clarity
- Improve IntegrationTestHelper documentation explaining usage contexts
